% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeData}
\alias{removeData}
\title{replace jims data to include NAs}
\usage{
removeData(sem, tsdata, seed, propMissing = 0.25, makePlot = F)
}
\arguments{
\item{sem}{sem. sem model definition}

\item{tsdata}{TS data frame. Assumes it has 8 columns (first being time) as in the vignette example}

\item{seed}{integer. random number generator seed}

\item{propMissing}{numeric. proportion of data to assign as NAs}

\item{makePlot}{Boolean. plot fitted sampled network}
}
\value{
list

\item{sampledTS}{new data set sampled with missing obs}
\item{fit}{fitted dsem model}
\item{plot}{as_fitted_DAG model}
}
\description{
replace jims data to include NAs
}
