% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimstest.r
\name{testJim}
\alias{testJim}
\title{Jims code from vignette
run example in vignette, then introduce missing value and compare fitted model}
\usage{
testJim()
}
\description{
Jims code from vignette
run example in vignette, then introduce missing value and compare fitted model
}
